requirements.txt
allure-pytest==2.9.45
pytest==7.1.2
pytest-xdist==2.5.0
requests==2.27.1
Faker==13.11.0